### 2022-06-26 #4510 @editorialbot
[PRE REVIEW]: Nanomesh: A Python workflow tool for generating meshes from image data
**Submitting author:** @stefsmeets (<a href="http://orcid.org/0000-0002-5413-9038">Stef Smeets</a>)
**Repository:** https://github.com/hpgem/nanomesh
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.9.0
