### 2022-08-01 #4510 @prashjha
Hello @jameshgrn, any chance you could review this submission for JOSS?
