### 2022-08-02 #4510 @prashjha
@jameshgrn, awesome. Generally speaking, reviewing software articles for JOSS entails testing the software/library and making sure software is well documented and general enough for broad usage. If you think something is not right with the software, for example, a test does not work or part of the code is not well documented, you can open an issue in the software repository and request authors to address these issues.

To make the review easy, you will be assigned a checklist designed to test various aspects of the open source library.

You can check out [this JOSS guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for carrying out reviews. You can also quickly look at one or two [past reviews in JOSS](https://github.com/openjournals/joss-reviews/issues?q=label%3Areview+is%3Aclosed+) to get the idea.

Let me know if you are still up for reviewing this submission and I will add you as a reviewer. Thanks.
