+++ paper/paper.bib
@@ -20,7 +20,7 @@
   url = {http://http://gmsh.info/},
   version = {4.6.0},
   date = {2020-06-22},
-} 
+}
 
 @book{cgal,
   title = {{CGAL} User and Reference Manual},
@@ -67,6 +67,8 @@ keywords = {mesh refinement, mesh quality, mesh improvement, boundary recovery, 
 }
 
 @InProceedings{triangle,
+doi="10.1007/bfb0014497",
+url="https://doi.org/10.1007/bfb0014497",
 author="Shewchuk, Jonathan Richard",
 editor="Lin, Ming C.
 and Manocha, Dinesh",
@@ -112,7 +114,9 @@ isbn="978-3-540-70680-9"
   volume={2},
   pages={e453},
   year={2014},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi={10.7717/peerj.453},
+  url={https://doi.org/10.7717/peerj.453}
 }
 
 @article{Hughes2005,
