### 2022-08-17 #4654 @jameshgrn
This is a very well written piece of software! congrats to the authors. The use case is well demonstrated and the software is very well documented. It's a nice interplay of computer vision, data engineering, and user functionality. I particularly enjoyed the inclusion of the metrics submodule, including metrics is something that you don't have to do, but makes your user's experience that much better, cheers. The one missing item from my checklist is the community guidelines section. Based on the journal requirements, my recommendation is **CONDITIONAL ACCEPT** after the authors add in a page or section about contribution guidelines. The Development Notes section is helpful but technical, a more accessible contributions section prefacing it is required, see the blurb from JOSS below

> #### There should be clear guidelines for third-parties wishing to:
>- Contribute to the software
>- Report issues or problems with the software
>- Seek support

I also strongly encourage the authors to consider how arguments are passed to their functions and to document them without referring to another webpage, commandline style arguments should be in the realm of CLIs, whereas readable arguments are the norm in python.

Congrats on the great piece of user-centered software.

#### PAPER

General Comments: I think the authors demonstrate need effectively and have delivered a well-written paper.

Specific Comments:
- L 48 Typo - "nananomesh"
- L 73 Typo - "guassian"

#### SOFTWARE
Specific Comments:
-the way _opts_ are passed seems somewhat cumbersome and not well documented (though documentation is linked). For example, Why not make 'generate_mesh()' take two separate arguments fro minimum angle and size of triangles? I'm sure this has been considered, but it seems to be not quite as 'pythonic' FWIW. On some further reading, I see it mimics the likely familiar commandline usage, but then maybe just have a CLI (or provide classic arguments and then your string format style)? Sorry to harp on this point but I think its important.
