### 2022-08-28 #4654 @vijaysm
The paper is well written. But I think it can be improved by fixing several spelling mistakes in the text and some missing punctuation.

Some specific comments:

- P2. L48: "Nananomesh" to "Nanomesh"
- P2. L60: "using the apply() method, which guarantees an object of the same time will be returned". Here do you mean  "same type" ?
- What is the "li" method? Is that an acronym? If there is a reference, please provide one.
- Is there a better way to provide thresholds in Nanomesh to capture finer features more accurately, and differentiate it from the background? I see the description about `max_edge_dist` in P3, but it is unclear from the text whether this will yield a finer resolution of the contour.
- Is the contour returned a discrete one? I see that Figure (2) shows that the triangulation is constrained on the contour boundary.
- P3. L84: "snipped" should be "snippet"
