+++ paper/paper.bib
@@ -148,3 +148,18 @@ isbn="978-3-540-70680-9"
   doi = {10.1103/PhysRevB.72.153102},
   url = {https://link.aps.org/doi/10.1103/PhysRevB.72.153102}
 }
+
+@article{Li1993,
+title = {Minimum cross entropy thresholding},
+journal = {Pattern Recognition},
+volume = {26},
+number = {4},
+pages = {617-625},
+year = {1993},
+issn = {0031-3203},
+doi = {https://doi.org/10.1016/0031-3203(93)90115-D},
+url = {https://www.sciencedirect.com/science/article/pii/003132039390115D},
+author = {C.H. Li and C.K. Lee},
+keywords = {Thresholding, Image segmentation, Minimum cross entropy, Maximum entropy method},
+abstract = {The threshold selection problem is solved by minimizing the cross entropy between the image and its segmented version. The cross entropy is formulated in a pixel-to-pixel basis between the two images and a computationally attractive algorithm employing the histogram is developed. Without making a priori assumptions about the population distribution, this method provides an unbiased estimate of a binarized version of the image in an information theoretic sense.}
+}
