### 2022-09-20 6c4ba3c
Update contributing guidelines (#288)
