+++ docs/contributing.rst
@@ -1,10 +1,12 @@
-# Contributing guidelines
+Contributing Guidelines
+=======================
 
 Welcome! *Nanomesh* is an open-source project that aims to streamline segmenting image data and generating 2D or 3D meshes. If you're trying *Nanomesh* with your data, your experience, questions, bugs you encountered, and suggestions for improvement are important to the success of the project.
 
-We have a [Code of Conduct](CODE_OF_CONDUCT.md), please follow it in all your interactions with the project.
+We have a `Code of Conduct <https://github.com/hpgem/nanomesh/blob/master/CODE_OF_CONDUCT.md>`__, please follow it in all your interactions with the project.
 
-## Submitting changes
+Submitting changes
+------------------
 
 When contributing to this repository, please first discuss the change you wish to make via an issue with the owners of this repository before making a change.
 
@@ -16,19 +18,20 @@ Contributions can come in the form of:
 - Updates to the documentation
 - ... ?
 
-We use the usual GitHub pull-request flow. For more info see [GitHub's own documentation](https://help.github.com/articles/using-pull-requests/).
+We use the usual GitHub pull-request flow. For more info see `GitHub's own documentation <https://help.github.com/articles/using-pull-requests/>`__.
 
 Typically this means:
 
-1. [Forking](https://docs.github.com/articles/about-forks) the repository and/or make a [new branch](https://docs.github.com/articles/about-branches)
+1. `Forking <https://docs.github.com/articles/about-forks>`__ the repository and/or make a `new branch <https://docs.github.com/articles/about-branches>`__
 2. Making your changes
 3. Make sure that the tests pass and add your own
 4. Update the documentation is updated for new features
 5. Pushing the code back to Github
-6. [Create a new Pull Request](https://help.github.com/articles/creating-a-pull-request/)
+6. `Create a new Pull Request <https://help.github.com/articles/creating-a-pull-request/>`__
 
 One of the code owners will review your code and request changes if needed. Once your changes have been approved, your contributions will become part of *duqtools*. 🎉
 
-## Development
+Development
+-----------
 
-Check out the [development notes](https://nanomesh.readthedocs.io/en/latest/development.html) to help get started with development.
+Check out the :ref:`development` to help get started with development.
