### 2022-09-20 #4654 @stefsmeets
@jameshgrn @vijaysm Thanks for reviewing Nanomesh and the feedback! I have addressed the spelling mistakes, clarified the text and update the contributing guidelines. I addressed the more in-depth comments and feedback below.

> The Development Notes section is helpful but technical, a more accessible contributions section prefacing it is required, see the blurb from JOSS below.

I agree. I have revised the contributing guidelines and made them available from the [documentation](https://nanomesh.readthedocs.io/en/latest/contributing.html) as well as the readme.

> Specific Comments: -the way _opts_ are passed seems somewhat cumbersome and not well documented (though documentation is linked). For example, Why not make 'generate_mesh()' take two separate arguments fro minimum angle and size of triangles? I'm sure this has been considered, but it seems to be not quite as 'pythonic' FWIW. On some further reading, I see it mimics the likely familiar commandline usage, but then maybe just have a CLI (or provide classic arguments and then your string format style)? Sorry to harp on this point but I think its important.

I agree that this is important. I have not been able to find the right interface for this, but not for a lack of trying. Many of the CLI parameters for `tetgen` and `triangle` are overloaded with different functions or are somehow coupled. This makes untangling this into a consistent API a bit of a rabbit hole that is difficult to get right. In the end we decided to stay as close to the well-documented CLI interface and spend our time elsewhere.

> Is there a better way to provide thresholds in Nanomesh to capture finer features more accurately, and differentiate it from the background? I see the description about max_edge_dist in P3, but it is unclear from the text whether this will yield a finer resolution of the contour.

I'm not aware of a better way to do this more accurately. Defining the threshold depends on the quality and contrast of the image. We expose the methods available in scikit-image that the user can choose from. Ultimately, the users know best how a feature is defined.

The `max_edge_dist` provides a way to tune the contour. Ideally the meshing algorithm decides the resolution of the contour based on the requested mesh size. Sometimes, a contour has a straight edge defined by (too) many points. This parameter redefines the edge to have points no more than x pixels apart.
Being able to tune this is necessary, because meshing algorithms will not remove input points. I updated the text to make this a bit more clear.

> Is the contour returned a discrete one? I see that Figure (2) shows that the triangulation is constrained on the contour boundary.

Yes, it is a discrete contour.
