### 2022-10-03 #4654 @prashjha
@stefsmeets could you please update the JOSS article to fix the above DOI error?
