+++ paper/paper.bib
@@ -157,7 +157,7 @@ number = {4},
 pages = {617-625},
 year = {1993},
 issn = {0031-3203},
-doi = {https://doi.org/10.1016/0031-3203(93)90115-D},
+doi = {10.1016/0031-3203(93)90115-D},
 url = {https://www.sciencedirect.com/science/article/pii/003132039390115D},
 author = {C.H. Li and C.K. Lee},
 keywords = {Thresholding, Image segmentation, Minimum cross entropy, Maximum entropy method},
