### 2022-10-03 #4654 @prashjha
Hi, @stefsmeets; while you fix the doi error, I will go over the JOSS paper. If there are suggestions on the draft, I will share them here. Once you have incorporated them, I will hand the paper to the editor-in-chief for a final decision.
