### 2022-10-04 #4654 @prashjha
Hi, @stefsmeets, see if these suggestions make sense for your draft:

- L18. `Many meshing libraries ...` in this sentence, you may want to say meshing libraries from images (??) so that the following sentence on not many software being available makes sense? Otherwise, for simple meshing given the definition of the domain, there are many open source software.
- L21. Should you have a comma after `primitive`?
- L32. Remove extra `the`.
- L43. '... in MeshContainers; this is ...'?
- L67. `Contour finding and`?
- L68. `Triangulation.`?
- L65. I think you should have a sentence something like `The code below creates contours around the features`(??) just above the code listing below L76.

After you have updated the draft, can you also do (if not done already) a 'tagged' release of your code and archive the release using zenodo or other methods? Make sure that the zenodo archive's title matches this JOSS submission's title.

Once you are done, I will hand your paper to EiC for the final decision.
