+++ paper/paper.md
@@ -31,9 +31,9 @@ bibliography: paper.bib
 
 # Statement of need
 
-Simulations based on finite element methods (FEM) often require the creation of a unstructured mesh that represents the topology and physical properties of the object under examination. Many meshing libraries exist and allow the creation of such meshes. However most of these tools are proprietary with sometimes a substantial fee due to the level of certification required by their application domains [@centaur; @scanip]. Some open source libraries do exist but often create meshes from a CAD design or well defined primitive [@gmsh, @cgal]. While these meshing libraries are invaluably useful for the study of idealized systems they do not allow the mesh to account for potential defects in the underlying topology of the object.
+Simulations based on finite element methods (FEM) often require the creation of a unstructured mesh that represents the topology and physical properties of the object under examination. Many meshing libraries exist and allow the creation of such meshes. Several of these are proprietary with sometimes a substantial fee due to the level of certification required by their application domains [@centaur; @scanip]. Some open source libraries do exist but often create meshes from a CAD design or well defined primitive [@gmsh; @cgal]. While these meshing libraries are invaluably useful for the study of idealized systems they do not allow the mesh to account for potential defects in the underlying topology of the object.
 
-For example, the calculation of the optical properties of nanocrystals is usually performed with an ideal nano-structure as substrate for the propagation of the Maxwell equations [@Koenderink2005; @Hughes2005]. Such simulations provide very valuable insight but ignore the effect that manufacturing imprecision of the nanometer-sized pores can have on the overall properties of the crystal. To resolve such structure-property relationship, meshes conforming to experimental images of real nanocrystals are needed. The subsequent simulation of wave propagation through these meshes using any FEM solver leads to a better understanding of the the impact that imperfections may have on the overall properties. Similar use cases in different fields of material science and beyond are expected. The direct FEM simulations on real device topology might bring very valuable insights. Through its user friendliness, code qualitiy, `nanomesh` will enable scientist running advanced simulations on meshes that accurately represent the devices that are manufactured experimentally.
+For example, the calculation of the optical properties of nanocrystals is usually performed with an ideal nano-structure as substrate for the propagation of the Maxwell equations [@Koenderink2005; @Hughes2005]. Such simulations provide very valuable insight but ignore the effect that manufacturing imprecision of the nanometer-sized pores can have on the overall properties of the crystal. To resolve such structure-property relationship, meshes conforming to experimental images of real nanocrystals are needed. The subsequent simulation of wave propagation through these meshes using any FEM solver leads to a better understanding of the impact that imperfections may have on the overall properties. Similar use cases in different fields of material science and beyond are expected. The direct FEM simulations on real device topology might bring very valuable insights. Through its user friendliness, code qualitiy, `nanomesh` will enable scientist running advanced simulations on meshes that accurately represent the devices that are manufactured experimentally.
 
 # Workflow and class hierarchy
 
@@ -41,7 +41,7 @@ A large part of the work of generating a mesh is to pre-process, filter, and seg
 
 \autoref{fig:flowchart} shows the `Nanomesh` workflow from left to right. The input data are read from a 2D or 3D `numpy` array [@numpy] into an `Image` object. `Nanomesh` has dedicated classes (`Mesher`s) to generate contours and triangulate or tetrahedralize the image data.
 
[…]
