+++ docs/conf.py
@@ -52,7 +52,7 @@ copyright = u'2020, '
 author = u'Stef Smeets'
 
 # The short X.Y version.
-version = release = '0.9.0'
+version = release = '0.9.1'
 
 # The language for content autogenerated by Sphinx.
 language = 'english'
