+++ setup.cfg
@@ -25,7 +25,7 @@ name = nanomesh
 project_urls =
     Bug Tracker = https://github.com/hpgem/nanomesh/issues
 url = https://github.com/hpgem/nanomesh
-version = 0.9.0
+version = 0.9.1
 
 [options]
 zip_safe = False
