### 2022-10-07 #4654 @stefsmeets
Hi @prashjha Thanks for the useful suggestions, I tried to incorporate all of them in the final draft.

I created a new release: https://github.com/hpgem/nanomesh/tree/0.9.1 and uploaded it to Zenodo: https://zenodo.org/record/7157382

Let me know if there if there are any remaining issues.
